LayoutTests/editing/selection/ios/bidi-visually-contiguous-selection-6.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true VisuallyContiguousBidiTextSelectionEnabled=true ] -->
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<meta name="viewport" content="width=device-width, initial-scale=1">
<meta charset="utf-8">
<style>
body, html {
    font-size: 20px;
    font-family: system-ui;
}
</style>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    description("Verifies that the selection does not unnecessarily clamp to LTR text when selecting the entire RTL paragraph from the start");

    target = document.getElementById("target");
    bounds = target.getBoundingClientRect();

    await UIHelper.longPressAtPoint(bounds.left + bounds.width - 10, bounds.top + bounds.height / 2);
    await UIHelper.ensurePresentationUpdate();
    await UIHelper.waitForSelectionToAppear();
    initialBounds = await UIHelper.selectionBounds();
    initialSelectedText = getSelection().toString();

    const {x, y} = UIHelper.midPointOfRect(await UIHelper.getSelectionEndGrabberViewRect());
    await UIHelper.sendEventStream(new UIHelper.EventStreamBuilder()
        .begin(x, y)
        .move(bounds.left - 1, y, 0.5)
        .end()
        .takeResult());
    await UIHelper.ensurePresentationUpdate();
    await UIHelper.waitForSelectionToAppear();
    finalSelectedText = getSelection().toString();
    finalBounds = await UIHelper.selectionBounds();

    shouldBe("initialBounds.left + initialBounds.width", "finalBounds.left + finalBounds.width");
    shouldBeEqualToString("initialSelectedText", "أُرسلت");
    shouldBeEqualToString("finalSelectedText", target.textContent);
    finishJSTest();
});
</script>
</head>
<body>
    <p dir="rtl"><span id="target">&#x202b;أُرسلت من الـ iPhone&#x202c;</span></p>
</body>
</html>
